// src/models/Enums/SqlCondition.ts
/*******************************************************************************
 * (C) Copyright
 * Logivations GmbH, Munich 2010-2022
 ******************************************************************************/

enum SqlCondition {
	AND = 'AND',
	OR = 'OR',
	AND_NOT = 'AND_NOT',
	OR_NOT = 'OR_NOT',
	NOT = '!',
}

export default SqlCondition;